requirements.txt
streamlit==1.29.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-google-genai>=0.0.3
google-generativeai==0.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
pinecone-client>=3.0.0
chromadb==0.4.22
tiktoken==0.5.2 